pregame hungerloss fixed

+++ src/main/kotlin/yt/lost/main/commands/pregame/JumpAndRunCommand.kt
@@ -25,7 +25,7 @@ class JumpAndRunCommand(private val plugin: KBingo, private val runningGame: Run
                 }
             }
 
-            val tmp = JumpAndRun((p0 as Player), (Math.random() * 100 + 100).roundToInt()+0.5.toDouble())
+            val tmp = JumpAndRun((p0 as Player), ((Math.random() * 100 + 100).roundToInt()+0.5))
             jr.add(tmp)
             plugin.server.pluginManager.registerEvents(tmp, plugin)
         }
